requirements.txt
numpy==1.26.3
streamlit==1.30.0
tensorflow==2.15.0.post1
fastapi
uvicorn
pillow
python-multipart
gdown